@@ -10,22 +10,29 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
+    "@reduxjs/toolkit": "^1.9.5",
     "dotenv": "^16.0.3",
     "express": "^4.18.2",
     "papaparse": "^5.4.1",
+    "prop-types": "^15.8.1",
     "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "react-csv": "^2.2.2",
+    "react-dom": "^18.2.0",
+    "react-redux": "^8.0.7"
   },
   "devDependencies": {
     "@babel/core": "^7.21.8",
     "@babel/preset-env": "^7.21.5",
     "@babel/preset-react": "^7.18.6",
+    "@babel/preset-typescript": "^7.21.5",
     "babel-loader": "^9.1.2",
     "css-loader": "^6.7.3",
     "html-webpack-plugin": "^5.5.1",
     "nodemon": "^2.0.22",
     "style-loader": "^3.3.2",
-    "webpack": "^5.82.1",
+    "ts-loader": "^9.4.3",
+    "typescript": "^5.1.3",
+    "webpack": "^5.85.0",
     "webpack-cli": "^5.1.1",
     "webpack-dev-server": "^4.15.0"
   }
